requirements.txt
Django==3.0.7
django-cors-headers==3.5.0
djangorestframework==3.11.0
psycopg2-binary==2.8.4
